pyproject.toml
[tool.poetry]
name = "pylambda-bot"
version = "0.2.0"
description = "Python lambda function for chatbot."
authors = ["yadach <yadach@outlook.jp>"]
license = "MIT"
readme = "README.md"
packages = [{include = "pylambda_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.6"
pyyaml = "^6.0"
requests = "^2.30.0"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pyproject-flake8 = "^6.0.0"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
exclude = ".venv"
ignore = "H102,H103,W503,E203"
max-line-length = 100

[tool.black]
line-length = 100
exclude = '''
(
    .pytest_cache
    | .venv
)
'''

[tool.isort]
include_trailing_comma = true
line_length = 100
sort_order = "native"
force_single_line = true
lines_between_types = 0
